Fix buffer manager failure false positive (#4221)
We currently reserve memory before evicting, and try to evict enough pages that the amount of used memory is no larger than the buffer pool, but having each thread do that independently means that each thread tries to evict enough memory for the total amount of memory being reserved concurrently. So if enough threads are evicting simultaneously and the number of evictable pages is small, we may evict sufficient memory for all threads, but not sufficient memory to get the free space below the required threshold (which was also higher than necessary since we were taking into account the current size being reserved twice on iterations after the first). If we instead allow a thread to finish reserving when they have evicted enough space for the memory they are trying to reserve regardless of the total used memory, this can be avoided.
There is still a possible race condition if a thread is unable to evict enough and hits the failure check before the other threads which evicted more than enough finish freeing the memory and updating the usedMemory counter, but the likelihood is much smaller and fixing that race condition will probably have a significant performance penalty.